fix: encode float/double/varint in correct format

+++ app/src/main/java/protobuf/magic/struct/Field.java
@@ -4,8 +4,8 @@ public record Field(int index, Type type, Node value, ByteRange byterange) {
   public Object parseValue() {
     return (Object)
         switch (type) {
-          case I32 -> value.asInteger();
-          case I64 -> value.asLong();
+          case I32 -> value.asFloat();
+          case I64 -> value.asDouble();
           case VARINT -> value.asLong();
           case LEN -> value.asString();
           default -> value.asBytes();
