+++ app/src/main/java/protobuf/magic/struct/VarintResult.java
@@ -1,5 +1,3 @@
 package protobuf.magic.struct;
 
-import java.math.BigInteger;
-
-public record VarintResult(BigInteger value, int length) {}
+public record VarintResult(Long value, int length) {}
